rabbit_feature_flags: New module to manage feature flags
This module provides the primary functions to query feature flags and to
enable/disable them.

There is a backend module, `rabbit_ff_registry` which is internal. This
module is "empty" in the repository. It's recompiled at runtime with all
the discovered feature flags. This way, querying feature flags is fast.

To find feature flags, the module walks through all loaded Erlang
applications and check their `rmq_feature_flags` environment key. It
should either contain a map of feature flags, or point to a module and
function to call to get that map.

[#159298729]